XRDDEV-746 Move components related to keys and certificates
