+++ src/proxy-ui-api/frontend/src/components/ui/ServiceIcon.vue
