+++ src/proxy-ui-api/frontend/src/router.ts
@@ -4,7 +4,7 @@ import Login from './views/Login.vue';
 import Base from './views/Base.vue';
 import Clients from '@/views/Clients/Clients.vue';
 import Client from '@/views/Clients/Client.vue';
-import Keys from './views/Keys.vue';
+import Keys from '@/views/KeysAndCertificates/Keys.vue';
 import Settings from '@/views/Settings/Settings.vue';
 import SystemParameters from '@/views/Settings/SystemParameters.vue';
 import BackupAndRestore from '@/views/Settings/BackupAndRestore.vue';
@@ -22,9 +22,9 @@ import LocalGroup from '@/views/LocalGroup/LocalGroup.vue';
 import Services from '@/views/Clients/Services.vue';
 import ServiceClients from '@/views/Clients/ServiceClients.vue';
 import ServiceDescriptionDetails from '@/views/ServiceDescriptionDetails.vue';
-import SignAndAuthKeys from '@/components/SignAndAuthKeys.vue';
-import SSTlsCertificate from '@/components/SSTlsCertificate.vue';
-import ApiKey from '@/components/ApiKey.vue';
+import SignAndAuthKeys from '@/views/KeysAndCertificates/SignAndAuthKeys.vue';
+import SSTlsCertificate from '@/views/KeysAndCertificates/SSTlsCertificate.vue';
+import ApiKey from '@/views/KeysAndCertificates/ApiKey.vue';
 import Token from '@/views/Token.vue';
 import Key from '@/views/Key.vue';
 import Certificate from '@/views/Certificate.vue';
