+++ src/proxy-ui-api/frontend/src/views/Clients/Services.vue
@@ -141,7 +141,7 @@ import AddWsdlDialog from './AddWsdlDialog.vue';
 import AddRestDialog from './AddRestDialog.vue';
 import DisableServiceDescDialog from './DisableServiceDescDialog.vue';
 import WarningDialog from '@/components/WarningDialog.vue';
-import ServiceIcon from '@/components/ServiceIcon.vue';
+import ServiceIcon from '@/components/ui/ServiceIcon.vue';
 
 import _ from 'lodash';
