+++ src/proxy-ui-api/frontend/src/views/KeysAndCertificates/ApiKey.vue
