+++ src/proxy-ui-api/frontend/src/views/KeysAndCertificates/TokenLoginDialog.vue
@@ -109,6 +109,6 @@ export default Vue.extend({
 </script>
 
 <style lang="scss" scoped>
-@import '../assets/dialogs';
+@import '../../assets/dialogs';
 </style>
